+++ application/src/main/java/fi/hsl/parkandride/core/domain/FacilityStatus.java
@@ -5,6 +5,8 @@ import javax.validation.constraints.NotNull;
 
 import org.joda.time.DateTime;
 
+import fi.hsl.parkandride.core.domain.validation.SpacesAvailableOrStatusRequired;
+
 @SpacesAvailableOrStatusRequired
 public class FacilityStatus {
     @NotNull
