+++ application/src/main/java/fi/hsl/parkandride/core/domain/validation/ElementLength.java
@@ -1,4 +1,4 @@
-package fi.hsl.parkandride.core.domain;
+package fi.hsl.parkandride.core.domain.validation;
 
 import static java.lang.annotation.ElementType.ANNOTATION_TYPE;
 import static java.lang.annotation.ElementType.CONSTRUCTOR;
@@ -15,12 +15,16 @@ import javax.validation.Constraint;
 import javax.validation.Payload;
 
 @Documented
-@Constraint(validatedBy = { SetElementNotBlankValidator.class })
+@Constraint(validatedBy = { SetElementLengthValidator.class })
 @Target({ METHOD, FIELD, ANNOTATION_TYPE, CONSTRUCTOR, PARAMETER })
 @Retention(RUNTIME)
-public @interface ElementNotBlank {
+public @interface ElementLength {
 
-    String message() default "{org.hibernate.validator.constraints.NotBlank.message}";
+    int min() default 0;
+
+    int max() default Integer.MAX_VALUE;
+
+    String message() default "{org.hibernate.validator.constraints.Length.message}";
 
     Class<?>[] groups() default { };
