+++ application/src/main/java/fi/hsl/parkandride/core/domain/validation/SpacesAvailableOrStatusRequired.java
@@ -1,4 +1,4 @@
-package fi.hsl.parkandride.core.domain;
+package fi.hsl.parkandride.core.domain.validation;
 
 import static java.lang.annotation.ElementType.TYPE;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
@@ -15,7 +15,7 @@ import javax.validation.Payload;
 @Target({ TYPE})
 @Retention(RUNTIME)
 public @interface SpacesAvailableOrStatusRequired {
-    String message() default "{fi.hsl.parkandride.core.domain.SpacesAvailableOrStatusRequired.message}";
+    String message() default "{fi.hsl.parkandride.core.domain.validation.SpacesAvailableOrStatusRequired.message}";
 
     Class<?>[] groups() default { };
